@@ -4,46 +4,46 @@
   "description": "homeautomation",
   "author": "florian hoffmann",
   "license": "MIT",
+  "sideEffects": false,
   "dependencies": {
     "axios": "^0.18.0",
     "babel-polyfill": "^6.26.0",
     "bluebird": "^3.5.1",
     "body-parser": "^1.18.3",
     "chart.js": "^2.7.2",
-    "docxtemplater": "3.7.0",
+    "docxtemplater": "3.8.3",
     "g": "^2.0.1",
     "int-to-rgb": "^1.0.0",
     "jszip": "2.6.1",
     "miio": "0.15.6",
-    "mongodb": "^3.1.1",
-    "npm": "^6.2.0",
-    "office-ui-fabric-react": "^6.35.0",
-    "react": "16.4.1",
-    "react-chartjs-2": "^2.7.3",
-    "react-dom": "16.4.1",
+    "mongodb": "^3.1.3",
+    "office-ui-fabric-react": "^6.50.1",
+    "react": "16.4.2",
+    "react-chartjs-2": "^2.7.4",
+    "react-dom": "16.4.2",
     "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
     "serve-favicon": "^2.5.0",
-    "typescript": "^2.9.2",
-    "webpack": "^4.16.0"
+    "typescript": "^3.0.1",
+    "webpack": "^4.16.5"
   },
   "devDependencies": {
     "@types/axios": "^0.14.0",
-    "@types/bluebird": "^3.5.22",
-    "@types/chart.js": "^2.7.27",
+    "@types/bluebird": "^3.5.23",
+    "@types/chart.js": "^2.7.31",
     "@types/express": "4.16.0",
     "@types/jszip": "3.1.4",
     "@types/microsoft-ajax": "0.0.33",
-    "@types/mongodb": "^3.1.1",
-    "@types/node": "^10.5.2",
-    "@types/react": "^16.4.6",
+    "@types/mongodb": "^3.1.4",
+    "@types/node": "^10.7.1",
+    "@types/react": "^16.4.11",
     "@types/react-chartjs-2": "^2.5.7",
-    "@types/react-dom": "16.0.6",
-    "@types/react-router": "^4.0.28",
-    "@types/react-router-dom": "^4.2.7",
+    "@types/react-dom": "16.0.7",
+    "@types/react-router": "^4.0.30",
+    "@types/react-router-dom": "^4.3.0",
     "@types/requirejs": "2.1.31",
     "@types/uuid": "^3.4.3",
-    "@types/webpack": "4.4.6",
+    "@types/webpack": "4.4.10",
     "awesome-typescript-loader": "^5.2.0",
     "babel-cli": "^6.26.0",
     "babel-core": "^6.26.3",
@@ -56,15 +56,15 @@
     "expose-loader": "^0.7.5",
     "express": "^4.16.3",
     "file-loader": "1.1.11",
-    "source-map-loader": "^0.2.3",
-    "style-loader": "^0.21.0",
-    "ts-loader": "4.4.2",
-    "tslint": "^5.10.0",
+    "source-map-loader": "^0.2.4",
+    "style-loader": "^0.22.1",
+    "ts-loader": "4.5.0",
+    "tslint": "^5.11.0",
     "tslint-loader": "^3.6.0",
-    "url-loader": "^1.0.1",
+    "url-loader": "^1.1.1",
     "uuid": "^3.3.2",
     "webpack-bundle-analyzer": "^2.13.1",
-    "webpack-cli": "^3.0.8",
+    "webpack-cli": "^3.1.0",
     "webpack-notifier": "^1.6.0"
   },
   "scripts": {
