@@ -3,19 +3,14 @@ let JSZip = require("jszip");
 let Docxtemplater = require("docxtemplater");
 
 let fs = require("fs");
-let path = require("path");
-
-const cfg = require("../../../config/config.json");
-declare let MONGO_DB_FILIALEN_COLLECTION_STRING: string;
-declare let MONGO_DB_ROUTEN_COLLECTION_STRING: string;
+// declare let MONGO_DB_FILIALEN_COLLECTION_STRING: string;
+// declare let MONGO_DB_ROUTEN_COLLECTION_STRING: string;
 
 class DokumentService {
   public CreateDocX(complateFilename: string, complateTemplateFilepath: string, templateMetaData: {}) {
     return new Promise(async (resolve, reject) => {
-
       // Load the docx file as a binary
-      let content = fs
-        .readFileSync(complateTemplateFilepath, "binary");
+      let content = fs.readFileSync(complateTemplateFilepath, "binary");
 
       let zip = new JSZip(content);
 
@@ -28,21 +23,19 @@ class DokumentService {
       try {
         // render the document (replace all occurences of {first_name} by John, {last_name} by Doe, ...)
         doc.render();
-      }
-      catch (error) {
+      } catch (error) {
         let e = {
           message: error.message,
           name: error.name,
           stack: error.stack,
-          properties: error.properties,
+          properties: error.properties
         };
         console.log(JSON.stringify({ error: e }));
         // The error thrown here contains additional information when logged with JSON.stringify (it contains a property object).
         throw error;
       }
 
-      let buf = doc.getZip()
-        .generate({ type: "nodebuffer" });
+      let buf = doc.getZip().generate({ type: "nodebuffer" });
 
       // buf is a nodejs buffer, you can either write it to a file or do anything else with it.
       fs.writeFileSync(complateFilename, buf);
