@@ -1,14 +1,7 @@
 import * as React from "react";
-import {
-  Toggle,
-  Slider,
-  Label,
-  Icon,
-  IconButton
-} from "office-ui-fabric-react";
+import { IconButton } from "office-ui-fabric-react";
 import { IBaseWeatherSensor } from "../../../interfaces/xiaomi";
 import { Panel } from "../../../global/components/simple/Panel";
-import Axios from "axios";
 import { Fragment } from "react";
 import { BaseWeatherSensorChart } from "./BaseWeatherSensorChart";
 
@@ -19,10 +12,7 @@ export interface IBaseWeatherSensorProps {
 export interface IBaseWeatherSensorState {
   showDetails: boolean;
 }
-export class BaseWeatherSensor extends React.Component<
-  IBaseWeatherSensorProps,
-  IBaseWeatherSensorState
-> {
+export class BaseWeatherSensor extends React.Component<IBaseWeatherSensorProps, IBaseWeatherSensorState> {
   constructor(props) {
     super(props);
     this.state = { showDetails: false };
@@ -58,36 +48,25 @@ export class BaseWeatherSensor extends React.Component<
               <Fragment>
                 <div className="ms-Grid-row">
                   <div className="ms-Grid-col ms-sm12">
-                    <h1 className="ms-font-su ms-fontColor-themePrimary">
-                      {this.props.sensorInformations.temperature + " °C"}
-                    </h1>
+                    <h1 className="ms-font-su ms-fontColor-themePrimary">{this.props.sensorInformations.temperature + " °C"}</h1>
                   </div>
                 </div>
                 <div className="ms-Grid-row">
                   <div className="ms-Grid-col ms-sm6">
                     <h1 className="ms-font-xl ms-fontColor-themePrimary">
                       {this.props.sensorInformations.humidity}
-                      <i
-                        className="ms-Icon ms-Icon--Precipitation"
-                        aria-hidden="true"
-                      />
+                      <i className="ms-Icon ms-Icon--Precipitation" aria-hidden="true" />
                     </h1>
                   </div>
                   {this.props.sensorInformations.hasPressure && (
                     <div className="ms-Grid-col ms-sm6">
-                      <h1 className="ms-font-xl ms-fontColor-themePrimary">
-                        {this.props.sensorInformations.pressure}
-                      </h1>
+                      <h1 className="ms-font-xl ms-fontColor-themePrimary">{this.props.sensorInformations.pressure}</h1>
                     </div>
                   )}
                 </div>
               </Fragment>
             )}
-            {this.state.showDetails && (
-              <BaseWeatherSensorChart
-                sensorInformations={this.props.sensorInformations}
-              />
-            )}
+            {this.state.showDetails && <BaseWeatherSensorChart sensorInformations={this.props.sensorInformations} />}
           </Panel>
         </div>
       </div>
