@@ -1,26 +1,15 @@
 import * as React from "react";
-import { Toggle } from "office-ui-fabric-react";
 import Axios from "axios";
-import {
-  ILightModel,
-  IRGBColor,
-  IBaseWeatherSensor,
-  IGatewayModel
-} from "../../../../interfaces/xiaomi";
-import { BaseLight } from "../../../../global/components/simple/BaseLight";
+import { IBaseWeatherSensor } from "../../../../interfaces/xiaomi";
 import { BaseWeatherSensor } from "../../../../global/components/simple/BaseWeatherSensor";
 import { setInterval } from "timers";
-const intToRGB = require("int-to-rgb");
 export interface IApplicationProps {}
 export interface IApplicationState {
   isInitialized: boolean;
   sensors: IBaseWeatherSensor[];
   intervalId: undefined;
 }
-export class Application extends React.Component<
-  IApplicationProps,
-  IApplicationState
-> {
+export class Application extends React.Component<IApplicationProps, IApplicationState> {
   private isMountedFinished = false;
   constructor(props) {
     super(props);
@@ -59,20 +48,13 @@ export class Application extends React.Component<
     return (
       <div className="ms-Grid-row">
         <div className="ms-Grid-col ms-sm12">
-          {(!this.state.sensors || this.state.sensors.length < 1) && (
-            <h1 className="ms-font-xl ms-fontColor-themePrimary">
-              Keine Sensoren gefunden
-            </h1>
-          )}
+          {(!this.state.sensors || this.state.sensors.length < 1) && <h1 className="ms-font-xl ms-fontColor-themePrimary">Keine Sensoren gefunden</h1>}
           <div className="ms-Grid-row">
             {this.state.sensors &&
               this.state.sensors.length > 0 &&
               this.state.sensors.map((sensor, index) => {
                 return (
-                  <div
-                    className="ms-Grid-col ms-sm12 ms-lg6 ms-xl3"
-                    key={"sensor_container_" + index}
-                  >
+                  <div className="ms-Grid-col ms-sm12 ms-lg6 ms-xl3" key={"sensor_container_" + index}>
                     <BaseWeatherSensor id={index} sensorInformations={sensor} />
                   </div>
                 );
