@@ -1,16 +1,9 @@
 import * as React from "react";
-import { Toggle } from "office-ui-fabric-react";
 import Axios from "axios";
-import {
-  ILightModel,
-  IRGBColor,
-  IBaseWeatherSensor,
-  IGatewayModel
-} from "../../../../interfaces/xiaomi";
+import { ILightModel, IRGBColor, IBaseWeatherSensor, IGatewayModel } from "../../../../interfaces/xiaomi";
 import { BaseLight } from "../../../../global/components/simple/BaseLight";
 import { BaseWeatherSensor } from "../../../../global/components/simple/BaseWeatherSensor";
 import { setInterval } from "timers";
-const intToRGB = require("int-to-rgb");
 export interface IApplicationProps {
   requestUrl: string;
 }
@@ -21,10 +14,7 @@ export interface IApplicationState {
   lights: ILightModel[];
   gatewayLights: ILightModel[];
 }
-export class Application extends React.Component<
-  IApplicationProps,
-  IApplicationState
-> {
+export class Application extends React.Component<IApplicationProps, IApplicationState> {
   constructor(props) {
     super(props);
     this.state = {
@@ -96,35 +86,19 @@ export class Application extends React.Component<
   }
   private colorChangedOnLight(lightInformation: ILightModel, color: IRGBColor) {
     let rgb = color.r * 65536 + color.g * 256 + color.b;
-    Axios.post("/api/lights/" + lightInformation.id + "/color/" + rgb).then(
-      this.reloadLightInformations
-    );
+    Axios.post("/api/lights/" + lightInformation.id + "/color/" + rgb).then(this.reloadLightInformations);
   }
   private powerChangedOnLight(lightInformation: ILightModel) {
-    Axios.post("/api/lights/" + lightInformation.id + "/power").then(
+    Axios.post("/api/lights/" + lightInformation.id + "/power").then(this.reloadLightInformations);
+  }
+  private colorSchemaChangedOnLight(lightInformation: ILightModel, color: IRGBColor, brightness: number) {
+    let rgb = color.r * 65536 + color.g * 256 + color.b;
+    Promise.all([Axios.post("/api/lights/" + lightInformation.id + "/brightness/" + brightness), Axios.post("/api/lights/" + lightInformation.id + "/color/" + rgb)]).then(
       this.reloadLightInformations
     );
   }
-  private colorSchemaChangedOnLight(
-    lightInformation: ILightModel,
-    color: IRGBColor,
-    brightness: number
-  ) {
-    let rgb = color.r * 65536 + color.g * 256 + color.b;
-    Promise.all([
-      Axios.post(
-        "/api/lights/" + lightInformation.id + "/brightness/" + brightness
-      ),
-      Axios.post("/api/lights/" + lightInformation.id + "/color/" + rgb)
-    ]).then(this.reloadLightInformations);
-  }
-  private brightnessChangedOnLight(
-    lightInformation: ILightModel,
-    brightness: number
-  ) {
-    Axios.post(
-      "/api/lights/" + lightInformation.id + "/brightness/" + brightness
-    ).then(this.reloadLightInformations);
+  private brightnessChangedOnLight(lightInformation: ILightModel, brightness: number) {
+    Axios.post("/api/lights/" + lightInformation.id + "/brightness/" + brightness).then(this.reloadLightInformations);
   }
   render() {
     if (!this.state.isInitialized) {
@@ -134,39 +108,25 @@ export class Application extends React.Component<
     return (
       <div className="ms-Grid-row">
         <div className="ms-Grid-col ms-sm12">
-          {(!this.state.sensors || this.state.sensors.length < 1) && (
-            <h1 className="ms-font-xl ms-fontColor-themePrimary">
-              Keine Sensoren gefunden
-            </h1>
-          )}
+          {(!this.state.sensors || this.state.sensors.length < 1) && <h1 className="ms-font-xl ms-fontColor-themePrimary">Keine Sensoren gefunden</h1>}
           <div className="ms-Grid-row">
             {this.state.sensors &&
               this.state.sensors.length > 0 &&
               this.state.sensors.map((sensor, index) => {
                 return (
-                  <div
-                    className="ms-Grid-col ms-sm12 ms-lg6 ms-xl3"
-                    key={"sensor_container_" + index}
-                  >
+                  <div className="ms-Grid-col ms-sm12 ms-lg6 ms-xl3" key={"sensor_container_" + index}>
                     <BaseWeatherSensor id={index} sensorInformations={sensor} />
                   </div>
                 );
               })}
           </div>
-          {(!this.state.lights || this.state.lights.length < 1) && (
-            <h1 className="ms-font-xl ms-fontColor-themePrimary">
-              Keine Lampen gefunden
-            </h1>
-          )}
+          {(!this.state.lights || this.state.lights.length < 1) && <h1 className="ms-font-xl ms-fontColor-themePrimary">Keine Lampen gefunden</h1>}
           <div className="ms-Grid-row">
             {this.state.lights &&
               this.state.lights.length > 0 &&
               this.state.lights.map((light, index) => {
                 return (
-                  <div
-                    className="ms-Grid-col ms-sm12 ms-lg6 ms-xl3"
-                    key={"light_container_" + index}
-                  >
+                  <div className="ms-Grid-col ms-sm12 ms-lg6 ms-xl3" key={"light_container_" + index}>
                     <BaseLight
                       lightInformation={light}
                       id={index}
@@ -179,66 +139,31 @@ export class Application extends React.Component<
                 );
               })}
           </div>
-          {(!this.state.gateways || this.state.gateways.length < 1) && (
-            <h1 className="ms-font-xl ms-fontColor-themePrimary">
-              Keine Gateways gefunden
-            </h1>
-          )}
+          {(!this.state.gateways || this.state.gateways.length < 1) && <h1 className="ms-font-xl ms-fontColor-themePrimary">Keine Gateways gefunden</h1>}
           <div className="ms-Grid-row">
             {this.state.gatewayLights &&
               this.state.gatewayLights.length > 0 &&
               this.state.gatewayLights.map((gw, index) => {
                 return (
-                  <div
-                    className="ms-Grid-col ms-sm12 ms-lg6 ms-xl3"
-                    key={"gwr_" + index}
-                  >
+                  <div className="ms-Grid-col ms-sm12 ms-lg6 ms-xl3" key={"gwr_" + index}>
                     <BaseLight
                       lightInformation={gw}
                       id={index}
-                      onBrightnessChanged={(
-                        lightInformation: ILightModel,
-                        brightness: number
-                      ) => {
-                        Axios.post(
-                          "/api/gateways/" +
-                            lightInformation.id +
-                            "/brightness/" +
-                            brightness
-                        ).then(this.reloadGatewayInformations);
+                      onBrightnessChanged={(lightInformation: ILightModel, brightness: number) => {
+                        Axios.post("/api/gateways/" + lightInformation.id + "/brightness/" + brightness).then(this.reloadGatewayInformations);
                       }}
-                      onColorChanged={(
-                        lightInformation: ILightModel,
-                        color: IRGBColor
-                      ) => {
-                        Axios.post(
-                          "/api/gateways/" + lightInformation.id + "/color",
-                          { color }
-                        ).then(this.reloadGatewayInformations);
+                      onColorChanged={(lightInformation: ILightModel, color: IRGBColor) => {
+                        Axios.post("/api/gateways/" + lightInformation.id + "/color", { color }).then(this.reloadGatewayInformations);
                       }}
-                      onColorSchemaChanged={(
-                        lightInformation: ILightModel,
-                        color: IRGBColor,
-                        brightness: number
-                      ) => {
-                        Axios.post(
-                          "/api/gateways/" + lightInformation.id + "/color",
-                          { color }
-                        )
+                      onColorSchemaChanged={(lightInformation: ILightModel, color: IRGBColor, brightness: number) => {
+                        Axios.post("/api/gateways/" + lightInformation.id + "/color", { color })
                           .then(() => {
-                            return Axios.post(
-                              "/api/gateways/" +
-                                lightInformation.id +
-                                "/brightness/" +
-                                brightness
-                            );
+                            return Axios.post("/api/gateways/" + lightInformation.id + "/brightness/" + brightness);
                           })
                           .then(this.reloadGatewayInformations);
                       }}
                       onPowerChanged={(lightInformation: ILightModel) => {
-                        Axios.post(
-                          "/api/gateways/" + lightInformation.id + "/power"
-                        ).then(this.reloadGatewayInformations);
+                        Axios.post("/api/gateways/" + lightInformation.id + "/power").then(this.reloadGatewayInformations);
                       }}
                     />
                   </div>
