@@ -1,28 +1,16 @@
 import * as React from "react";
-import { Toggle } from "office-ui-fabric-react";
 import Axios from "axios";
-import {
-  ILightModel,
-  IRGBColor,
-  IBaseWeatherSensor,
-  IGatewayModel
-} from "../../../../interfaces/xiaomi";
-import { BaseLight } from "../../../../global/components/simple/BaseLight";
-import { BaseWeatherSensor } from "../../../../global/components/simple/BaseWeatherSensor";
+import { ILightModel, IRGBColor } from "../../../../interfaces/xiaomi";
 import { setInterval } from "timers";
 import { Yeelight } from "../simple/Yeelight";
 import { Fragment } from "react";
-const intToRGB = require("int-to-rgb");
 export interface IApplicationProps {}
 export interface IApplicationState {
   isInitialized: boolean;
   lights: ILightModel[];
   intervalId: undefined;
 }
-export class Application extends React.Component<
-  IApplicationProps,
-  IApplicationState
-> {
+export class Application extends React.Component<IApplicationProps, IApplicationState> {
   private isMountedFinished = false;
   constructor(props) {
     super(props);
@@ -32,9 +20,7 @@ export class Application extends React.Component<
     this.powerChangedOnLight = this.powerChangedOnLight.bind(this);
     this.colorSchemaChangedOnLight = this.colorSchemaChangedOnLight.bind(this);
     this.brightnessChangedOnLight = this.brightnessChangedOnLight.bind(this);
-    this.colorTemperatureChangedOnLight = this.colorTemperatureChangedOnLight.bind(
-      this
-    );
+    this.colorTemperatureChangedOnLight = this.colorTemperatureChangedOnLight.bind(this);
     this.reloadLightInformations = this.reloadLightInformations.bind(this);
     this.loadDevices = this.loadDevices.bind(this);
   }
@@ -73,43 +59,22 @@ export class Application extends React.Component<
   }
   private colorChangedOnLight(lightInformation: ILightModel, color: IRGBColor) {
     let rgb = color.r * 65536 + color.g * 256 + color.b;
-    Axios.post("/api/lights/" + lightInformation.id + "/color/" + rgb).then(
-      this.reloadLightInformations
-    );
+    Axios.post("/api/lights/" + lightInformation.id + "/color/" + rgb).then(this.reloadLightInformations);
   }
   private powerChangedOnLight(lightInformation: ILightModel) {
-    Axios.post("/api/lights/" + lightInformation.id + "/power").then(
+    Axios.post("/api/lights/" + lightInformation.id + "/power").then(this.reloadLightInformations);
+  }
+  private colorTemperatureChangedOnLight(lightInformation: ILightModel, colorTemperature: number) {
+    Axios.post("/api/lights/" + lightInformation.id + "/temperature/" + colorTemperature).then(this.reloadLightInformations);
+  }
+  private colorSchemaChangedOnLight(lightInformation: ILightModel, color: IRGBColor, brightness: number) {
+    let rgb = color.r * 65536 + color.g * 256 + color.b;
+    Promise.all([Axios.post("/api/lights/" + lightInformation.id + "/brightness/" + brightness), Axios.post("/api/lights/" + lightInformation.id + "/color/" + rgb)]).then(
       this.reloadLightInformations
     );
   }
-  private colorTemperatureChangedOnLight(
-    lightInformation: ILightModel,
-    colorTemperature: number
-  ) {
-    Axios.post(
-      "/api/lights/" + lightInformation.id + "/temperature/" + colorTemperature
-    ).then(this.reloadLightInformations);
-  }
-  private colorSchemaChangedOnLight(
-    lightInformation: ILightModel,
-    color: IRGBColor,
-    brightness: number
-  ) {
-    let rgb = color.r * 65536 + color.g * 256 + color.b;
-    Promise.all([
-      Axios.post(
-        "/api/lights/" + lightInformation.id + "/brightness/" + brightness
-      ),
-      Axios.post("/api/lights/" + lightInformation.id + "/color/" + rgb)
-    ]).then(this.reloadLightInformations);
-  }
-  private brightnessChangedOnLight(
-    lightInformation: ILightModel,
-    brightness: number
-  ) {
-    Axios.post(
-      "/api/lights/" + lightInformation.id + "/brightness/" + brightness
-    ).then(this.reloadLightInformations);
+  private brightnessChangedOnLight(lightInformation: ILightModel, brightness: number) {
+    Axios.post("/api/lights/" + lightInformation.id + "/brightness/" + brightness).then(this.reloadLightInformations);
   }
   render() {
     if (!this.state.isInitialized) {
@@ -119,20 +84,13 @@ export class Application extends React.Component<
     return (
       <div className="ms-Grid-row">
         <div className="ms-Grid-col ms-sm12">
-          {(!this.state.lights || this.state.lights.length < 1) && (
-            <h1 className="ms-font-xl ms-fontColor-themePrimary">
-              Keine Lampen gefunden
-            </h1>
-          )}
+          {(!this.state.lights || this.state.lights.length < 1) && <h1 className="ms-font-xl ms-fontColor-themePrimary">Keine Lampen gefunden</h1>}
           <div className="ms-Grid-row">
             {this.state.lights &&
               this.state.lights.length > 0 &&
               this.state.lights.map((light, index) => {
                 return (
-                  <div
-                    className="ms-Grid-col ms-sm12 ms-lg6 ms-xl3"
-                    key={"light_container_" + index}
-                  >
+                  <div className="ms-Grid-col ms-sm12 ms-lg6 ms-xl3" key={"light_container_" + index}>
                     <Fragment>
                       {/* <BaseLight
                         lightInformation={light}
@@ -149,9 +107,7 @@ export class Application extends React.Component<
                         onColorChanged={this.colorChangedOnLight}
                         onColorSchemaChanged={this.colorSchemaChangedOnLight}
                         onPowerChanged={this.powerChangedOnLight}
-                        onColorTemperatureChanged={
-                          this.colorTemperatureChangedOnLight
-                        }
+                        onColorTemperatureChanged={this.colorTemperatureChangedOnLight}
                       />
                     </Fragment>
                   </div>
